requirements.txt
mysql
Flask==1.1.2
Flask-Cors==3.0.10
